Add outline and fix sticker issues
