+++ Assets/Scripts/Sticker/ToolSticker.cs
@@ -1,12 +1,11 @@
-using System;
 using UnityEngine;
 
 public class ToolSticker : Sticker
 {
-    [SerializeField] float distance = 10;
     Transform target;
     Vector2 tolerance  = new Vector2(125, 160);
     [SerializeField] protected GameObject prefab;
+    bool removed = false;
 
     protected override void Initialize() {
         enabled = true;
@@ -17,7 +16,6 @@ public class ToolSticker : Sticker
 
     protected override void UseStickerBefore()
     {
-
         for (int i = 0; i < 2; i++) {
             if (Mathf.Abs(transform.position[i] - target.position[i]) > tolerance[i]) {
                 transform.localPosition = Vector3.zero;
@@ -25,6 +23,7 @@ public class ToolSticker : Sticker
             }
         }
         Instantiate(prefab).transform.position = TransformationFunction.BookToWorld(target.localPosition);
+        inventoryDisplay.RemoveSticker(this);
         Destroy(gameObject);
     }
 }
